requirements.txt
click==8.1.3
Flask==2.0.2
gunicorn==20.1.0
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing-extensions==4.7.1
Werkzeug==2.0.2
zipp==3.6.0
flask_cors==3.0.10
mysql-connector-python==8.0.26
flask_jwt_extended
bcrypt
Pillow==9.0.1
requests